Add missing `post-delete` finalizers to ArgoCD app
The Helm chart has some post-delete hooks, so the ArgoCD application
controller sets the `post-delete` finalizers. We add those in the App
definition to avoid ArgoCD constantly removing and adding those
finalizers.

We need to do this because the ArgoCD component library always sets the
`resources-finalizer.argocd.argoproj.io` finalizer.